requirements.txt
openai==2.12.0
streamlit==1.52.1
python-dotenv==1.2.1
trafilatura==2.0.0
tenacity==9.1.2
numpy==2.3.5
requests==2.32.5
